Unified the way in which a default avatar is recovered in many cases

+++ app/src/main/java/mega/privacy/android/app/lollipop/megachat/chatAdapters/MegaChatLollipopAdapter.java
@@ -24,7 +24,6 @@ import android.text.Spannable;
 import android.text.SpannableString;
 import android.text.Spanned;
 import android.text.TextUtils;
-import android.text.format.Formatter;
 import android.text.style.ForegroundColorSpan;
 import android.text.style.RelativeSizeSpan;
 import android.text.style.StyleSpan;
@@ -639,7 +638,6 @@ public class MegaChatLollipopAdapter extends RecyclerView.Adapter<RecyclerView.V
         RoundedImageView contentOwnMessageContactThumb;
         private EmojiTextView contentOwnMessageContactName;
         public EmojiTextView contentOwnMessageContactEmail;
-        private EmojiTextView contentOwnMessageContactInitialLetter;
         RelativeLayout forwardOwnContact;
 
         ImageView iconOwnTypeDocLandPreview;
@@ -684,8 +682,6 @@ public class MegaChatLollipopAdapter extends RecyclerView.Adapter<RecyclerView.V
         private EmojiTextView nameContactText;
 
         RoundedImageView contactImageView;
-        private EmojiTextView contactInitialLetter;
-
         RelativeLayout contentContactMessageLayout;
         EmojiTextView contentContactMessageText;
 
@@ -756,7 +752,6 @@ public class MegaChatLollipopAdapter extends RecyclerView.Adapter<RecyclerView.V
 
         public void setMyImageView(Bitmap bitmap) {
             contentOwnMessageContactThumb.setImageBitmap(bitmap);
-            contentOwnMessageContactInitialLetter.setVisibility(View.GONE);
         }
 
         public void setContactImageView(Bitmap bitmap) {
@@ -967,7 +962,6 @@ public class MegaChatLollipopAdapter extends RecyclerView.Adapter<RecyclerView.V
             holder.contentOwnMessageContactName = v.findViewById(R.id.content_own_message_contact_name);
             holder.contentOwnMessageContactEmail = v.findViewById(R.id.content_own_message_contact_email);
 
-            holder.contentOwnMessageContactInitialLetter = v.findViewById(R.id.content_own_message_contact_initial_letter);
             holder.forwardOwnContact = v.findViewById(R.id.forward_own_contact);
             holder.forwardOwnContact.setTag(holder);
             holder.forwardOwnContact.setVisibility(View.GONE);
@@ -1013,8 +1007,6 @@ public class MegaChatLollipopAdapter extends RecyclerView.Adapter<RecyclerView.V
             holder.titleContactMessage = v.findViewById(R.id.title_contact_message_layout);
 
             holder.contactImageView = v.findViewById(R.id.contact_thumbnail);
-            holder.contactInitialLetter = v.findViewById(R.id.contact_initial_letter);
-
             holder.timeContactText = v.findViewById(R.id.contact_message_chat_time_text);
             holder.nameContactText = v.findViewById(R.id.contact_message_chat_name_text);
 
@@ -5772,7 +5764,9 @@ public class MegaChatLollipopAdapter extends RecyclerView.Adapter<RecyclerView.V
                 holder.contentOwnMessageContactEmail.setText(name);
                 String email = context.getResources().getQuantityString(R.plurals.general_selection_num_contacts, (int) userCount, userCount);
                 holder.contentOwnMessageContactName.setText(email);
-                createDefaultAvatar(holder, null, email, true, userCount);
+                int color = colorAvatar(context, megaApi, -1);
+                Bitmap bitmapDefaultAvatar = getDefaultAvatar(context, color, userCount + "", AVATAR_SIZE, true);
+                holder.contentOwnMessageContactThumb.setImageBitmap(bitmapDefaultAvatar);
             }
 
             if (!multipleSelect) {
@@ -5888,7 +5882,9 @@ public class MegaChatLollipopAdapter extends RecyclerView.Adapter<RecyclerView.V
                 holder.contentContactMessageContactEmail.setText(name);
                 String email = context.getResources().getQuantityString(R.plurals.general_selection_num_contacts, (int) userCount, userCount);
                 holder.contentContactMessageContactName.setText(email);
-                createDefaultAvatar(holder, null, email, false, userCount);
+                int color = colorAvatar(context, megaApi, -1);
+                Bitmap bitmap = getDefaultAvatar(context, color, userCount + "", AVATAR_SIZE, true);
+                holder.contentContactMessageContactThumb.setImageBitmap(bitmap);
             }
 
             if (!multipleSelect) {
@@ -6639,20 +6635,19 @@ public class MegaChatLollipopAdapter extends RecyclerView.Adapter<RecyclerView.V
     public void setUserAvatar(ViewHolderMessageChat holder, MegaChatMessage message) {
         logDebug("setUserAvatar");
 
-        String name = "";
-        name = message.getUserName(0);
+        String name = message.getUserName(0);
         if (name.trim().isEmpty()) {
             name = message.getUserEmail(0);
         }
         String email = message.getUserEmail(0);
-        String userHandleEncoded = MegaApiAndroid.userHandleToBase64(message.getUserHandle(0));
-
         ChatAttachmentAvatarListener listener;
+        int color = colorAvatar(context, megaApi, message.getUserHandle());
+        Bitmap bitmapDefaultAvatar = getDefaultAvatar(context, color, name, AVATAR_SIZE, true);
         if (myUserHandle == message.getUserHandle()) {
-            createDefaultAvatar(holder, userHandleEncoded, name, true,0);
+            holder.contentOwnMessageContactThumb.setImageBitmap(bitmapDefaultAvatar);
             listener = new ChatAttachmentAvatarListener(context, holder, this, true);
         } else {
-            createDefaultAvatar(holder, userHandleEncoded, name, false, 0);
+            holder.contentContactMessageContactThumb.setImageBitmap(bitmapDefaultAvatar);
             listener = new ChatAttachmentAvatarListener(context, holder, this, false);
         }
 
@@ -6675,7 +6670,6 @@ public class MegaChatLollipopAdapter extends RecyclerView.Adapter<RecyclerView.V
                     megaApi.getUserAvatar(email,buildAvatarFile(context,email + ".jpg").getAbsolutePath(),listener);
                 } else {
                     if (myUserHandle == message.getUserHandle()) {
-                        holder.contentOwnMessageContactInitialLetter.setVisibility(View.GONE);
                         holder.contentOwnMessageContactThumb.setImageBitmap(bitmap);
                     } else {
                         holder.contentContactMessageContactInitialLetter.setVisibility(View.GONE);
@@ -6700,21 +6694,18 @@ public class MegaChatLollipopAdapter extends RecyclerView.Adapter<RecyclerView.V
         }
     }
 
-    public void setContactAvatar(ViewHolderMessageChat holder, long userHandle, String name) {
-        logDebug("setContactAvatar");
-
+    private void setContactAvatar(ViewHolderMessageChat holder, long userHandle, String name) {
+        /*Default Avatar*/
         String userHandleEncoded = MegaApiAndroid.userHandleToBase64(userHandle);
-        createDefaultAvatarContact(holder, userHandleEncoded, name);
+        int color = colorAvatar(context, megaApi, userHandle);
+        holder.contactImageView.setImageBitmap(getDefaultAvatar(context, color, name, AVATAR_SIZE, true));
+
+        /*Avatar*/
         ChatAttachmentAvatarListener listener = new ChatAttachmentAvatarListener(context, holder, this, false);
-
         File avatar = buildAvatarFile(context, userHandleEncoded + ".jpg");
-
-        Bitmap bitmap = null;
-        if (isFileAvailable(avatar)) {
-            if (avatar.length() > 0) {
+        Bitmap bitmap;
+        if (isFileAvailable(avatar) && avatar.length() > 0) {
                 BitmapFactory.Options bOpts = new BitmapFactory.Options();
-                bOpts.inPurgeable = true;
-                bOpts.inInputShareable = true;
                 bitmap = BitmapFactory.decodeFile(avatar.getAbsolutePath(), bOpts);
                 if (bitmap == null) {
                     avatar.delete();
@@ -6726,17 +6717,8 @@ public class MegaChatLollipopAdapter extends RecyclerView.Adapter<RecyclerView.V
 
                     megaApi.getUserAvatar(userHandleEncoded, buildAvatarFile(context, userHandleEncoded + ".jpg").getAbsolutePath(), listener);
                 } else {
-                    holder.contactInitialLetter.setVisibility(View.GONE);
                     holder.contactImageView.setImageBitmap(bitmap);
                 }
-            } else {
-                if (megaApi == null) {
-                    logWarning("megaApi is Null in Offline mode");
-                    return;
-                }
-
-                megaApi.getUserAvatar(userHandleEncoded, buildAvatarFile(context, userHandleEncoded + ".jpg").getAbsolutePath(), listener);
-            }
         } else {
 
             if (megaApi == null) {
@@ -6745,218 +6727,6 @@ public class MegaChatLollipopAdapter extends RecyclerView.Adapter<RecyclerView.V
             }
 
             megaApi.getUserAvatar(userHandleEncoded, buildAvatarFile(context, userHandleEncoded + ".jpg").getAbsolutePath(), listener);
-        }
-    }
-
-    public void createDefaultAvatarContact(ViewHolderMessageChat holder, String userHandle, String name) {
-        logDebug("createDefaultAvatarContact");
-        Bitmap defaultAvatar = Bitmap.createBitmap(DEFAULT_AVATAR_WIDTH_HEIGHT, DEFAULT_AVATAR_WIDTH_HEIGHT, Bitmap.Config.ARGB_8888);
-        Canvas c = new Canvas(defaultAvatar);
-        Paint p = new Paint();
-        p.setAntiAlias(true);
-
-        String color = megaApi.getUserAvatarColor(userHandle);
-        if (userHandle != null) {
-            if (color != null) {
-                logDebug("The color to set the avatar is " + color);
-                p.setColor(Color.parseColor(color));
-            } else {
-                logDebug("Default color to the avatar");
-                p.setColor(ContextCompat.getColor(context, R.color.lollipop_primary_color));
-            }
-        } else {
-            p.setColor(ContextCompat.getColor(context, R.color.lollipop_primary_color));
-        }
-
-        int radius;
-        if (defaultAvatar.getWidth() < defaultAvatar.getHeight())
-            radius = defaultAvatar.getWidth() / 2;
-        else
-            radius = defaultAvatar.getHeight() / 2;
-
-        c.drawCircle(defaultAvatar.getWidth() / 2, defaultAvatar.getHeight() / 2, radius, p);
-
-        Display display = ((Activity) context).getWindowManager().getDefaultDisplay();
-        outMetrics = new DisplayMetrics();
-        display.getMetrics(outMetrics);
-
-        holder.contactImageView.setImageBitmap(defaultAvatar);
-
-        if (name != null) {
-            if (name.trim().length() > 0) {
-                String firstLetter = getFirstLetter(name);
-                if(firstLetter.trim().isEmpty() || firstLetter.equals("(")){
-                    holder.contactInitialLetter.setVisibility(View.INVISIBLE);
-                }else {
-                    holder.contactInitialLetter.setText(firstLetter);
-                    holder.contactInitialLetter.setTextColor(Color.WHITE);
-                    holder.contactInitialLetter.setVisibility(View.VISIBLE);
-                }
-            }
-        }
-    }
-
-    public void createGroupChatAvatar(ViewHolderMessageChat holder, String title, boolean isMyMsg){
-        logDebug("createGroupChatAvatar()");
-
-        Bitmap defaultAvatar = Bitmap.createBitmap(DEFAULT_AVATAR_WIDTH_HEIGHT,DEFAULT_AVATAR_WIDTH_HEIGHT, Bitmap.Config.ARGB_8888);
-        Canvas c = new Canvas(defaultAvatar);
-        Paint p = new Paint();
-        p.setAntiAlias(true);
-        p.setColor(ContextCompat.getColor(context,R.color.divider_upgrade_account));
-
-        int radius;
-        if (defaultAvatar.getWidth() < defaultAvatar.getHeight())
-            radius = defaultAvatar.getWidth()/2;
-        else
-            radius = defaultAvatar.getHeight()/2;
-
-        c.drawCircle(defaultAvatar.getWidth()/2, defaultAvatar.getHeight()/2, radius, p);
-
-        if(isMyMsg){
-            holder.urlOwnMessageGroupAvatar.setImageBitmap(defaultAvatar);
-        }
-        else{
-            holder.urlContactMessageGroupAvatar.setImageBitmap(defaultAvatar);
-        }
-
-        Display display = ((Activity)context).getWindowManager().getDefaultDisplay();
-        outMetrics = new DisplayMetrics ();
-        display.getMetrics(outMetrics);
-        float density  = context.getResources().getDisplayMetrics().density;
-
-        if (title.length() > 0){
-            String chatTitle = title.trim();
-            String firstLetter = chatTitle.charAt(0) + "";
-            firstLetter = firstLetter.toUpperCase(Locale.getDefault());
-            if(firstLetter.trim().isEmpty()){
-                if(isMyMsg){
-                    holder.urlOwnMessageGroupAvatarText.setVisibility(View.INVISIBLE);
-                }
-                else{
-                    holder.urlContactMessageGroupAvatarText.setVisibility(View.INVISIBLE);
-                }
-            }
-            else{
-                if(firstLetter.equals("(")){
-                    if(isMyMsg){
-                        holder.urlOwnMessageGroupAvatarText.setVisibility(View.INVISIBLE);
-                    }
-                    else{
-                        holder.urlContactMessageGroupAvatarText.setVisibility(View.INVISIBLE);
-                    }
-                }
-                else{
-                    if(isMyMsg){
-                        holder.urlOwnMessageGroupAvatarText.setText(firstLetter);
-                        holder.urlOwnMessageGroupAvatarText.setTextColor(Color.WHITE);
-                        holder.urlOwnMessageGroupAvatarText.setVisibility(View.VISIBLE);
-                        holder.urlOwnMessageGroupAvatarText.setTextSize(32);
-                    }
-                    else{
-                        holder.urlContactMessageGroupAvatarText.setText(firstLetter);
-                        holder.urlContactMessageGroupAvatarText.setTextColor(Color.WHITE);
-                        holder.urlContactMessageGroupAvatarText.setVisibility(View.VISIBLE);
-                        holder.urlContactMessageGroupAvatarText.setTextSize(32);
-                    }
-                }
-            }
-        }
-    }
-
-    public void createDefaultAvatar(ViewHolderMessageChat holder, String userHandle, String name, boolean isMyMsg, long userCount) {
-        logDebug("createDefaultAvatar()");
-
-        Bitmap defaultAvatar = Bitmap.createBitmap(DEFAULT_AVATAR_WIDTH_HEIGHT, DEFAULT_AVATAR_WIDTH_HEIGHT, Bitmap.Config.ARGB_8888);
-        Canvas c = new Canvas(defaultAvatar);
-        Paint p = new Paint();
-        p.setAntiAlias(true);
-
-        String color = megaApi.getUserAvatarColor(userHandle);
-        if (userHandle != null) {
-            if (color != null) {
-                logDebug("The color to set the avatar is " + color);
-                p.setColor(Color.parseColor(color));
-            } else {
-                logDebug("Default color to the avatar");
-                p.setColor(ContextCompat.getColor(context, R.color.lollipop_primary_color));
-            }
-        } else {
-            p.setColor(ContextCompat.getColor(context, R.color.lollipop_primary_color));
-        }
-
-
-        int radius;
-        if (defaultAvatar.getWidth() < defaultAvatar.getHeight())
-            radius = defaultAvatar.getWidth() / 2;
-        else
-            radius = defaultAvatar.getHeight() / 2;
-
-        c.drawCircle(defaultAvatar.getWidth() / 2, defaultAvatar.getHeight() / 2, radius, p);
-
-        Display display = ((Activity) context).getWindowManager().getDefaultDisplay();
-        outMetrics = new DisplayMetrics();
-        display.getMetrics(outMetrics);
-
-        if (isMyMsg) {
-            holder.contentOwnMessageContactThumb.setImageBitmap(defaultAvatar);
-            if(userHandle!=null){
-                if (name != null) {
-                    if (name.trim().length() > 0) {
-
-                        String firstLetter = getFirstLetter(name);
-                        if(firstLetter.trim().isEmpty() || firstLetter.equals("(")){
-                            holder.contentOwnMessageContactInitialLetter.setVisibility(View.INVISIBLE);
-                        }else {
-                            holder.contentOwnMessageContactInitialLetter.setText(firstLetter);
-                            holder.contentOwnMessageContactInitialLetter.setTextColor(Color.WHITE);
-                            holder.contentOwnMessageContactInitialLetter.setVisibility(View.VISIBLE);
-                        }
-
-                    }
-                    holder.contentOwnMessageContactInitialLetter.setTextSize(24);
-
-                }
-            }else{
-                if (name != null) {
-                    String firstLetter = userCount + "";
-                    holder.contentOwnMessageContactInitialLetter.setText(firstLetter);
-                    holder.contentOwnMessageContactInitialLetter.setTextColor(Color.WHITE);
-                    holder.contentOwnMessageContactInitialLetter.setVisibility(View.VISIBLE);
-                    holder.contentOwnMessageContactInitialLetter.setTextSize(24);
-                }
-            }
-
-        } else {
-            holder.contentContactMessageContactThumb.setImageBitmap(defaultAvatar);
-
-            if(userHandle!=null){
-                if (name != null) {
-                    if (name.trim().length() > 0) {
-
-                        String firstLetter = getFirstLetter(name);
-                        if(firstLetter.trim().isEmpty() || firstLetter.equals("(")){
-                            holder.contentContactMessageContactInitialLetter.setVisibility(View.INVISIBLE);
-                        }else {
-                            holder.contentContactMessageContactInitialLetter.setText(firstLetter);
-                            holder.contentContactMessageContactInitialLetter.setTextColor(Color.WHITE);
-                            holder.contentContactMessageContactInitialLetter.setVisibility(View.VISIBLE);
-                        }
-                    }
-                    holder.contentContactMessageContactInitialLetter.setTextSize(24);
-
-                }
-            }else{
-                if (name != null) {
-                    String firstLetter = userCount + "";
-                    holder.contentContactMessageContactInitialLetter.setText(firstLetter);
-                    holder.contentContactMessageContactInitialLetter.setTextColor(Color.WHITE);
-                    holder.contentContactMessageContactInitialLetter.setVisibility(View.VISIBLE);
-                    holder.contentContactMessageContactInitialLetter.setTextSize(24);
-
-                }
-            }
-
         }
     }
